requirement.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
websockets>=11.0
requests>=2.31.0
aiofiles>=23.1.0